Use #mw-head, not #mw-navigation, for navigation link
mw-navigation is a collapsed div because all of its children
are absolutely positioned, and it's in a totally useless
place. Instead, point to #mw-head, which is where we pointed
to originally before it was changed to #mw-navigation.

This reverts 75858d59d from when this code was still in MediaWiki core.

Bug: T78207
Change-Id: I5e1f0edbf80f6f5f330df73cb06e4b0ee4323905