Make an empty revision for 1.23
Vector was part of MediaWiki core in 1.23. Create an empty
directory with a no-op entry point for the REL1_23 branch so that
things that check skins out automatically based on branch (such as
MW-Vagrant) fail more gracefully. Also helps people who are trying
to find the right version via git log.

Change-Id: I1474d95461cd43d1bd6299339f3e4310794629f3