Make vectorMenu expand as tabs in IE6
Right now, Vector loads the csshover.htc extension in order to
emulate CSS :hover support in IE6.

This commit makes the vectorMenu expand as basic tabs instead.
Thus allowing us to remove the csshover.htc extension.

Note that the "page action" vector menu (p-cactions) is only created if it
contains items. By default, the only actions in this menu are actions for
logged-in users. For logged-out users, the menu isn't even on the page.

The tabs aren't pretty, but they don't look broken either.

Change-Id: Ibf2332ab0a0d9d7ba3ec03665ad61ca287750ad3