collapsibleTabs: Stop the tabs from collapsing back and forth forever
Update the 'expandedWidth' after we expand in case someone was brazen
enough to change the tab's contents after the page load (gasp).

This doesn't prevent a tab from collapsing back and forth once, but at
least it won't continue to do that forever. This should reduce the
issue from completely awful to mildly annoying.

Bug: T71729
Change-Id: I46bd6584c0f2ddebc4aa7d1103cff1715e1510a5