Use makeListItem() for menu items rather than building HTML by hand
This affects the four tabs menus (namespaces, variants, views, actions).
Other menus (personal menu and sidebar) have already been using it.

We need some minor overrides to get the same results as the hand-built
HTML, but I think this still makes the code a lot nicer.

The output is the same as before, except for unimportant whitespace
differences and the order of some tag attributes.

I tested this with several extensions and configuration options that
mess with the tabs:
* $wgUsePigLatinVariant = true
* VisualEditor extension
* FileAnnotations extension
* FileExporter extension
* Viewing the page as administrator

Change-Id: I2d1255442abf5fa4bac2de1b084d0bcacbba7d0f