Print styles: set minimum width for paragraphs
An arbitrary 120pt (approved by the designer) has been chosen as the
minimum width of paragraphs. This is especially useful when a paragraph
is surrounded by two floating elements from the left and right. If there
is not enough space in the middle, then the change will push the
paragraph text until there's enough space to display the text.

Bug: T175008
Change-Id: I428f153309162e23606e5187a0e1bd7244253f10