OOUI: Remove confusing 'position' override
Vector is the only skin that does this, setting 'position: relative'
for .oo-ui-windowManager-modal (normally 'position: static') and for
.ve-ui-overlay-global (normally 'position: absolute').

The override for .oo-ui-windowManager-modal caused it to create a new
stacking context, which is the only reason the 'z-index' override
worked. Use the right selector to override 'z-index' instead.

The override for .ve-ui-overlay-global was completely pointless and
it's surprising that it never broke anything.

Change-Id: Icd1dec43e2da9ef2090b18145099838de3a7890a